Refactor folders & Add next-auth
