+++ .env.example
@@ -1,1 +1,2 @@
 NEXT_PUBLIC_API_URL=HTTP Address 
+NEXT_AUTH_SECRET=secret
