+++ app/dashboard/casos/page.tsx
@@ -5,9 +5,9 @@ import { useRouter } from 'next/navigation';
 import { Flex, Text, Box, Button } from '@chakra-ui/react';
 import { AddIcon, EditIcon, ViewIcon } from '@chakra-ui/icons';
 
-import { DataItem, MainActionButtonProps } from '@/app/types';
-import { ObjectTable } from '@/app/components/object-table';
-import { ObjectTabs } from '@/app/components/object-tabs';
+import { DataItem, MainActionButtonProps } from '@/types/types';
+import { ObjectTable } from '@/components/object-table';
+import { ObjectTabs } from '@/components/object-tabs';
 
 const tabs: string[] = ['Todos'];
