+++ app/layout.tsx
@@ -1,28 +1,21 @@
-'use client';
-
-import { CacheProvider } from '@chakra-ui/next-js';
-import { ChakraProvider } from '@chakra-ui/react';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { useRef } from 'react';
+import { getServerSession } from 'next-auth';
+import { authConfig } from '@/pages/api/auth/[...nextauth]';
+import { Providers } from '@/components/providers';
 
 import './globals.css';
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const queryClient = useRef(new QueryClient());
+  const session = await getServerSession(authConfig);
 
   return (
-    <html lang="en">
+    <html lang="es">
       <head />
       <body>
-        <QueryClientProvider client={queryClient.current}>
-          <CacheProvider>
-            <ChakraProvider>{children}</ChakraProvider>
-          </CacheProvider>
-        </QueryClientProvider>
+        <Providers session={session}>{children}</Providers>
       </body>
     </html>
   );
