--- app/login/useLogin.ts
@@ -1,7 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-
-import { login } from '@/networking/services/login';
-
-export const useLogin = () => {
-  return useMutation(login);
-};
