+++ components/object-tabs/index.tsx
