+++ package.json
@@ -16,8 +16,8 @@
     "@chakra-ui/react": "^2.5.1",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
-    "@tanstack/react-query": "^4.26.1",
     "@hookform/resolvers": "^3.0.0",
+    "@tanstack/react-query": "^4.26.1",
     "@types/node": "18.15.2",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
@@ -26,6 +26,7 @@
     "eslint-config-next": "13.2.4",
     "framer-motion": "^10.2.5",
     "next": "13.2.4",
+    "next-auth": "^4.20.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.7",
