+++ utils/constants.ts
@@ -1,0 +1,1 @@
+export const DASHBOARD_HOME_PATH = '/dashboard';
